+++ src/main/java/com/Cccccc/airbnb/domain/entity/Facility.java
@@ -1,4 +1,4 @@
-package com.Cccccc.airbnb.domain.user.entity;
+package com.Cccccc.airbnb.domain.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,12 +11,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserIntroCategory {
+public class Facility {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private String category_name;
+    private String facilityName;
 
-    private String category_image;
+    private String facilityImage;
+
+    private Integer mainType;
+
+    private String subType;
 }
