+++ src/main/java/com/Cccccc/airbnb/domain/entity/RoomCategory.java
@@ -1,4 +1,4 @@
-package com.Cccccc.airbnb.domain.user.entity;
+package com.Cccccc.airbnb.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -8,18 +8,17 @@ import lombok.NoArgsConstructor;
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserIntro {
+public class RoomCategory {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "category_id",nullable = false)
+    private Category category;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_intro_category_id",nullable = false)
-    private UserIntroCategory userIntroCategory;
-
-    private String content;
+    @JoinColumn(name = "room_id", nullable = false)
+    private Room room;
 }
