+++ src/main/java/com/Cccccc/airbnb/domain/entity/UserIntro.java
@@ -1,4 +1,4 @@
-package com.Cccccc.airbnb.domain.room.entity;
+package com.Cccccc.airbnb.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -8,17 +8,18 @@ import lombok.NoArgsConstructor;
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RoomCategory {
-
+public class UserIntro {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "category_id",nullable = false)
-    private Category category;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "room_id", nullable = false)
-    private Room room;
+    @JoinColumn(name = "user_intro_category_id",nullable = false)
+    private UserIntroCategory userIntroCategory;
+
+    private String content;
 }
