+++ src/main/java/com/Cccccc/airbnb/domain/room/exception/RoomNotFoundException.java
@@ -1,4 +1,4 @@
-package com.Cccccc.airbnb.domain.room.controller.common;
+package com.Cccccc.airbnb.domain.room.exception;
 
 public class RoomNotFoundException extends RuntimeException {
     public RoomNotFoundException(String message) {
