+++ src/main/java/com/Cccccc/airbnb/domain/room/repository/RoomRepository.java
@@ -1,6 +1,6 @@
 package com.Cccccc.airbnb.domain.room.repository;
 
-import com.Cccccc.airbnb.domain.room.entity.Room;
+import com.Cccccc.airbnb.domain.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository extends JpaRepository<Room, Integer> {
