+++ src/main/java/com/Cccccc/airbnb/domain/room/service/RoomService.java
@@ -1,17 +1,17 @@
-package com.Cccccc.airbnb.domain.room.service.host;
+package com.Cccccc.airbnb.domain.room.service;
 
 
-import com.Cccccc.airbnb.domain.room.controller.common.RoomNotFoundException;
-import com.Cccccc.airbnb.domain.room.controller.host.dto.HostResponseDto;
-import com.Cccccc.airbnb.domain.room.entity.Room;
+import com.Cccccc.airbnb.domain.room.exception.RoomNotFoundException;
+import com.Cccccc.airbnb.domain.room.dto.HostResponseDto;
+import com.Cccccc.airbnb.domain.entity.Room;
 import com.Cccccc.airbnb.domain.room.repository.RoomRepository;
-import com.Cccccc.airbnb.domain.user.entity.User;
+import com.Cccccc.airbnb.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class HostService {
+public class RoomService {
     private final RoomRepository  RoomRepository ;
 
     public HostResponseDto getUserByRoomId(Integer roomId) {
